requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
pandas==2.2.2
numpy==1.26.4
whitenoise
django-widget-tweaks==1.4.12
dj-database-url
psycopg[binary]